Tried to fix the login.
Added the user and header so that it is saved to file. The buttons works
and the api works. It get the user from the api. User is presented in
the navigation drawer. Fragments gets started correctly (for the login).
Problem: crash when writing or reading (something wrong with either of
them).