Network control, Fixes.
Added controls so that it does not crash when no internet is present.
Made so that the search fragment now works. The title detail fragment
also works. Changed the looks of both (change back if pref other). Moved
the styles of the title detail to file. SearchItem press gets title
detail page for now.