Borrowings fragment
Added real icons, replacing the placeholders. There are special icons
for when a loan is recalled and when a reservation has a loanable
available. Also the background for those items are set. Date is shown
without time. Toast when checking in. TODO: Set TextView when no items
are available / no internet.